refactor: WinningLogic.java 제거, WinningLogic 메소드를 Prizes와 WinningLotto로 이동
